Change int64 back to int32

+++ src/weathergen/datasets/tokenizer_forecast.py
@@ -263,7 +263,7 @@ class TokenizerForecast:
         target_times_raw = np.split(times_reordered, ll)
         target_times = np.split(times_reordered_enc, ll)
 
-        target_tokens_lens = torch.tensor([len(s) for s in target_tokens], dtype=torch.int64)
+        target_tokens_lens = torch.tensor([len(s) for s in target_tokens], dtype=torch.int32)
 
         # compute encoding of target coordinates used in prediction network
         if target_tokens_lens.sum() > 0:
